slamd-96: Created new ExperimentPostprocessor class
